list/xl: Fix the way marker is handled in leafDirectory verification.
