+++ web-router.go
@@ -30,7 +30,7 @@ import (
 
 // webAPI container for Web API.
 type webAPIHandlers struct {
-	ObjectAPI *objectAPI
+	ObjectAPI objectAPI
 }
 
 // indexHandler - Handler to serve index.html
